v12.2.1 fix service total nights to acuumulate with float

+++ public/components/scoutbookServiceActivity.js
@@ -11,7 +11,7 @@ function ScoutbookServiceActiviy($scope) {
         var nights=0;
         if (this.scout && this.scout._activities && this.scout._activities._service) {
             for (i=0;i<this.scout._activities._service.length;i++) {
-                var amount = parseInt(this.scout._activities._service[i]._count);
+                var amount = parseFloat(this.scout._activities._service[i]._count);
                 nights += amount;
             }
         }
